message_to_score/pyproject.toml
[project]
name = "message-to-score"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.29.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
]
